bundle_add: Move bundle_add specific functions to bundle_add.c
